+++ extensions-utils/src/dependencies/download_wasms/nns.rs
@@ -21,7 +21,7 @@ pub async fn download_nns_wasms(dfx_cache_path: &Path) -> anyhow::Result<()> {
         wasm_name,
         test_wasm_name,
         ..
-    } in NNS_CORE
+    } in [NNS_CORE.to_vec(), NNS_CORE_MANUAL.to_vec()].concat()
     {
         download_ic_repo_wasm(wasm_name, &ic_commit, wasm_dir).await?;
         if let Some(test_wasm_name) = test_wasm_name {
@@ -127,6 +127,27 @@ pub const NNS_IC_CKBTC_MINTER: IcNnsInitCanister = IcNnsInitCanister {
     test_wasm_name: None,
     canister_id: "qjdve-lqaaa-aaaaa-aaaeq-cai",
 };
+/// ICRC1 ledger (used for ckETH).
+pub const ICRC1_LEDGER: IcNnsInitCanister = IcNnsInitCanister {
+    canister_name: "nns-icrc1-ledger",
+    wasm_name: "ic-icrc1-ledger.wasm",
+    test_wasm_name: None,
+    canister_id: "qvhpv-4qaaa-aaaaa-aaagq-cai",
+};
+/// ICRC1 index (used for ckETH).
+pub const ICRC1_INDEX: IcNnsInitCanister = IcNnsInitCanister {
+    canister_name: "nns-icrc1-index-ng",
+    wasm_name: "ic-icrc1-index-ng.wasm",
+    test_wasm_name: None,
+    canister_id: "q4eej-kyaaa-aaaaa-aaaha-cai",
+};
+/// ICP index.
+pub const ICP_INDEX: IcNnsInitCanister = IcNnsInitCanister {
+    canister_name: "nns-icp-index",
+    wasm_name: "ic-icp-index-canister.wasm",
+    test_wasm_name: None,
+    canister_id: "q3fc5-haaaa-aaaaa-aaahq-cai",
+};
 /// Minimum data needed to download and deploy a standard canister via dfx deploy NAME.
 pub struct StandardCanister {
     /// The typical name of the canister, as seen in dfx.json or used in `dfx canister id NAME`.
@@ -143,14 +164,21 @@ pub const INTERNET_IDENTITY: StandardCanister = StandardCanister {
     canister_name: "internet_identity",
     canister_id: "qhbym-qaaaa-aaaaa-aaafq-cai",
     wasm_name: "internet_identity_dev.wasm",
-    wasm_url: "https://github.com/dfinity/internet-identity/releases/download/release-2022-07-11/internet_identity_dev.wasm"
+    wasm_url: "https://github.com/dfinity/internet-identity/releases/download/release-2024-05-13/internet_identity_dev.wasm.gz"
 };
 /// Frontend dapp for voting and managing neurons.
 pub const NNS_DAPP: StandardCanister = StandardCanister {
     canister_name: "nns-dapp",
     canister_id: "qsgjb-riaaa-aaaaa-aaaga-cai",
-    wasm_name: "nns-dapp_local.wasm",
-    wasm_url: "https://github.com/dfinity/nns-dapp/releases/download/tip/nns-dapp_t2.wasm",
+    wasm_name: "nns-dapp_test.wasm",
+    wasm_url: "https://github.com/dfinity/nns-dapp/releases/download/proposal-129748/nns-dapp_test.wasm.gz",
+};
+/// SNS aggregator canister.
+pub const SNS_AGGREGATOR: StandardCanister = StandardCanister {
+    canister_name: "sns-aggregator",
+    canister_id: "sgymv-uiaaa-aaaaa-aaaia-cai",
+    wasm_name: "sns_aggregator.wasm.gz",
+    wasm_url: "https://github.com/dfinity/nns-dapp/releases/download/proposal-129748/sns_aggregator.wasm.gz",
 };
 /// Canister that can be used instead of cycle wallets
 pub const CYCLES_LEDGER: StandardCanister = StandardCanister {
@@ -159,24 +187,27 @@ pub const CYCLES_LEDGER: StandardCanister = StandardCanister {
     wasm_name: "cycles_ledger.wasm.gz",
     wasm_url: "https://github.com/dfinity/cycles-ledger/releases/download/cycles-ledger-v0.2.8/cycles-ledger.wasm.gz",
 };
-/// Backend canisters deployed by `ic nns init`.
+/// Backend canisters deployed by `ic-nns-init`.
 pub const NNS_CORE: &[&IcNnsInitCanister; 11] = &[
-    &NNS_REGISTRY,       // 0
-    &NNS_GOVERNANCE,     // 1
-    &NNS_LEDGER,         // 2
-    &NNS_ROOT,           // 3
-    &NNS_CYCLES_MINTING, // 4
-    &NNS_LIFELINE,       // 5
-    &NNS_GENESIS_TOKENS, // 6
-    &NNS_IDENTITY,       // 7
-    &NNS_UI,             // 8
-    // 9 - unused
-    &NNS_SNS_WASM, // 10
-    // ic-icrc1-ledger is expected to go in place 11.
-    &NNS_IC_CKBTC_MINTER, // Index not defined here: https://github.com/dfinity/ic/blob/master/rs/nns/constants/src/lib.rs
+    &NNS_REGISTRY,        // 0
+    &NNS_GOVERNANCE,      // 1
+    &NNS_LEDGER,          // 2
+    &NNS_ROOT,            // 3
+    &NNS_CYCLES_MINTING,  // 4
+    &NNS_LIFELINE,        // 5
+    &NNS_GENESIS_TOKENS,  // 6
+    &NNS_IDENTITY,        // 7
+    &NNS_UI,              // 8
+    &NNS_IC_CKBTC_MINTER, // 9 - only in local deployment
+    &NNS_SNS_WASM,        // 10
 ];
-/// Frontend canisters deployed by `ic nns init`.  The deployment is normal, like any other canister.
-pub const NNS_FRONTEND: [&StandardCanister; 2] = [&INTERNET_IDENTITY, &NNS_DAPP];
+pub const NNS_CORE_MANUAL: &[&IcNnsInitCanister; 3] = &[
+    &ICRC1_LEDGER, // 13 - only in local deployment
+    &ICRC1_INDEX,  // 14 - only in local deployment
+    &ICP_INDEX,    // 15 - only in local deployment
+];
+/// Frontend canisters are not deployed by `ic-nns-init`. The deployment is normal, like any other canister.
+pub const NNS_FRONTEND: [&StandardCanister; 3] = [&INTERNET_IDENTITY, &NNS_DAPP, &SNS_AGGREGATOR];
 
 /// Test account with well known public & private keys, used in NNS_LEDGER, NNS_DAPP and third party projects.
 /// The keys use the ED25519 curve, used for BasicIdentity on th eInternet Computer.