+++ extensions/nns/CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## [Unreleased] - ReleaseDate
 
+## [0.4.1] - 2024-05-29
+- Bump II and NNS dapp to their latest mainnet verions (II: release-2024-05-13; NNS-Dapp: proposal-129748) and install their dependencies (ICRC1 ledger and index for ckETH, ICP index, SNS aggregator).
+
 ## [0.4.0] - 2024-04-04
 - Same functionality as version `0.3.1`.
 
@@ -29,6 +32,7 @@
 
 <!-- next-url -->
 [Unreleased]: https://github.com/dfinity/dfx-extensions/compare/{{tag_name}}...HEAD
+[0.4.1]: https://github.com/dfinity/dfx-extensions/compare/{{tag_name}}...{{tag_name}}
 [0.4.0]: https://github.com/dfinity/dfx-extensions/compare/nns-v0.3.1...{{tag_name}}
 [0.3.1]: https://github.com/dfinity/dfx-extensions/compare/nns-v0.3.0...nns-v0.3.1
 [0.3.0]: https://github.com/dfinity/dfx-extensions/compare/nns-v0.2.1...nns-v0.3.0